Fixed bug with fractionalpolaraxis use
The problem was python finding an old mpl_toolkits directory before the
one that lives in matplotlib-1.5.1*.egg, so there was a version mismatch
between matplotlib and axisartists. The solution is totally lame:
explicately add matplotlib*.egg/mpl_toolkits to pythonpath and import
axisartist (and others) directly. This is dumb and ensures this module
will only work on the uw astro computers. Grrrrrr.